build: migrate util.js to TypeScript
